refactor: extract large registry methods to RegistryOperations
- Extract _ensure_service_device_exists (~541 lines) to registry.py
- Extract _find_tracker_entity_entry and find_tracker_entity_entry (~281 lines) to registry.py
- Add required imports: Sequence, SimpleNamespace, EntityRegistryEntry, TRACKER_SUBENTRY_KEY, DEVICE_TRACKER_DOMAIN
- Add helper function imports for entity matching and unique_id building
- main.py reduced from 7,654 to 6,832 lines
- registry.py increased from 409 to 1,256 lines
